feat(poll): 투표 상태를 보여주는 컴포넌트 생성
